requirements.txt
flask
flask-socketio
eventlet
mysql-connector-python
pyzmq
eventlet
gunicorn
Flask-Session
